# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AthletOS API",
    "description": "A governor for athletes who already train.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/enrollments/{id}/exercise-adjustments": {
      "get": {
        "tags": [
          "enrollments"
        ],
        "summary": "Reads adjustments for an enrolment owned by the authenticated athlete.\nClosed enrolments remain readable so their settings do not disappear from\nhistory and settings screens merely because the block ended.",
        "operationId": "show_exercise_adjustments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The enrolment's id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The enrolment's exercise adjustments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseAdjustments"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such enrolment belongs to this athlete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "put": {
        "tags": [
          "enrollments"
        ],
        "summary": "Replaces every adjustment for one active enrolment.",
        "description": "The request first lands as generic JSON so a fractional number can be\ntranslated into this API's RFC 9457 422 response. Extracting the DTO\ndirectly would let Axum own that rejection instead of `ApiError`.",
        "operationId": "replace_exercise_adjustments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The enrolment's id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceExerciseAdjustments"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The adjustments as now stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseAdjustments"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such enrolment belongs to this athlete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The enrolment is closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "An exercise or percentage is not adjustable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExerciseAdjustments": {
        "type": "object",
        "description": "The canonical non-zero adjustments stored for one enrolment.",
        "required": [
          "enrollment_id",
          "adjustments"
        ],
        "properties": {
          "adjustments": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "propertyNames": {
              "type": "string"
            },
            "example": {
              "bench": 10,
              "squat": -5
            }
          },
          "enrollment_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 problem details. `type` is omitted until we publish stable problem\ntype URIs; consumers should branch on `status` plus `title` for now.",
        "required": [
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type.",
            "example": "Not Found"
          }
        }
      },
...
      "ReplaceExerciseAdjustments": {
        "type": "object",
        "description": "The complete adjustment document supplied by a client.",
        "required": [
          "adjustments"
        ],
        "properties": {
          "adjustments": {
            "type": "object",
            "description": "Integer percentages keyed by exercise. Zero removes an adjustment.",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "propertyNames": {
              "type": "string"
            },
            "example": {
              "bench": 10,
              "squat": -5
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519 (EdDSA) access token from /v1/auth/login, verifiable against /.well-known/jwks.json"
      }
    }
  },
...
}